Refactor: 単一ページのDomainのディレクトリを作成
