--- app/domains/auth/logout_domain.rb
@@ -1,9 +1,0 @@
-class Auth::LogoutDomain
-  def initialize(cookies)
-    @cookies = cookies
-  end
-
-  def execute
-    @cookies.delete(:jwt)
-  end
-end
