--- app/domains/auth/set_current_user_domain.rb
@@ -1,15 +1,0 @@
-class Auth::SetCurrentUserDomain
-  def initialize(token)
-    @token = token
-  end
-
-  def execute
-    return nil unless @token
-
-    decoded = JsonWebToken.decode(@token)
-    User.find_by(id: decoded[:user_id]) if decoded
-  rescue JWT::DecodeError => e
-    Rails.logger.warn("JWT::DecodeError occurred: #{e.message}")
-    nil
-  end
-end
