--- app/domains/auth/signup_domain.rb
@@ -1,16 +1,0 @@
-class Auth::SignupDomain
-  def initialize(user_params)
-    super()
-    @user_params = user_params
-  end
-
-  def execute
-    user = User.new(@user_params)
-    if user.save
-      token = JsonWebToken.encode(user_id: user.id)
-      [token, user]
-    else
-      raise user.errors.full_messages.join(", ")
-    end
-  end
-end
