+++ app/domains/page/auth_page/logout_domain.rb
@@ -1,0 +1,12 @@
+module Page::AuthPage
+  class LogoutDomain < ApplicationDomain
+    def initialize(cookies)
+      super
+      @cookies = cookies
+    end
+
+    def execute
+      @cookies.delete(:jwt)
+    end
+  end
+end
