+++ app/domains/page/profile_page/show_user_with_profile_domain.rb
@@ -1,6 +1,7 @@
-module Users
-  class ShowUserWithProfileDomain
+module Page::ProfilePage
+  class ShowUserWithProfileDomain < ApplicationDomain
     def initialize(user_id:)
+      super
       @user_id = user_id
     end
