--- app/domains/posts/show_post_domain.rb
@@ -1,12 +1,0 @@
-module Posts
-  class ShowPostDomain
-    def initialize(id:)
-      @id = id
-    end
-
-    def execute
-      post = Post.includes(:user, :replies, :likes, :reposts).find(@id)
-      PostDetailDto.new(post)
-    end
-  end
-end
