+++ app/services/jwt_token_service.rb
@@ -1,4 +1,4 @@
-class JwtTokenService
+class JwtTokenService < ApplicationService
   def self.encode(user)
     JsonWebToken.encode(user_id: user.id)
   end
